Visuals for 3D Calib Marker
If you want to take a look> open the PupilGaze>>Calibration>>3D and
play. there is no 3D calibration methos attached to the visuals yet, but
you can test how it looks by pressing the "A" button on the keyboard for
now. Every press represents one step in the samples of the calibration.