3D Calib Marker update + failsafes
I have advanced with the 3D Calib Marker. Now it is reacting to the
calibrationsampling and it does switch position for each calibration
point. Unfortunately I could not include the GUI of the calibraiton
point handling yet as planned, but it is coming in the next update! I
have handeled a potential freezing issue that we have faced when the
calibration process was terminated mid-process. I have also included a
check for the Calibration button in the GUI. Now it is disabled if there
is no connection established.