Hotfix for 2D and 3D calibration
2D and 3D calibration did not switch to processing gaze after
calibration was completed. This was due to a function that had to be
queued to the mainthread (GameObject methods can only de called from the
mainthread). A MainThread script is included to handle queuing
processes. Please note that when setting the Calibraiton point positions
there are some restrictions: 2D all coordinates must be in normalized
space 0.0-1.0. 3D , currently must have at least 100 units on the Z
axis. This might change later when Im going to match the units to cm/m
and to Perspecive.