Debug View + Image Data + GUI update
- Receiving Image data from  pupil service and rendering it in the scene
on the far plane of the camera frustum when Debug View is enabled
currently only under Calibration>>3D>>Calibration Debug Mode(Toggle)
fov, nearplane(minViewDistance), farplane(maxViewDistance) of camera
frustum is currently fixed to a public variable that you can access by
turning the DEV option on. (Should be around the bottom). The rendered
mesh for the eyes, is currently still a built in Sphere that is
referenced by a public var too : Debug Eye Mesh, you can find it in the
same place. (this will change in further versions, probably will include
a custom mesh).
A GUI change was made under the Settings>>Service. The Service IP is no
longer accessable to write when switched to Local (toggles back to the
default 127.0.0.1 and disables the fields). Debug View eye positioning
sill needs to be tested.