Debug View update + Operator Monitor
Operator Monitor button added to the Main Menu option. This creates a
camera that will only render (currently it does render on both) on the
screen and not on the screens of the HMD device. A Graph has been added
to track the confidence levels of the eyes. This graph is adjustable
through the variables exposed by the DEV options. (scaling, position,
rotation, updatespeed, amount of samples to show at once).
For the Debug View, the calculation of the fov has been changed to match
how Unity's. Also a camera frustum has been added to the Calibration
Points editor to match the points with the actual view.
Some GUI changes has also been applied : Under the Calibration Points
list the Vector 2 and Vector 3 - s were displayed with a great distance
between the point lable and the value, it is corrected now.
Toast messages functionality added