More Adjustments
- normal lenght is supposed to be picked up by the current normal from
the gaze data scaled by the magnitude of the gaze position 3D, yet to be
fixed!
- frustum distance set to have the 3D eyes on the camera image,
currently hardcoded to a distance of 30 unity units, that is equivalent
to a best setup if the HTC Vive eye distance from lenses is physically
set to minimum. Yet to implement clamped distance change on eyecenter
position refresh
- coordinate system fixed, yet to be confirmed
- change check added for the eyeball center positions. currently firing
off the method : OnUpdateEyeCenter on the main thread ( cannot access
some functions in the LineDrawer on secondary threads )
- Eye process activity check moved to pupil data, now at start we
subscribe to pupil data, initialize and indicate camera activity and
unsubscribe from pupil to reduce bandwidth usage.
- On Operator monitor start and stop we are now subscribing to, and
unsubscribing from the pupil data, and is the source for the confidence
level shown on the operator monitor graph.
- eye camera frustums set to 90 degrees.
Known issues:
-ffmpeg implementation is not yet compatible with interactive and VR.
-still no fluent way to be able to use the operator monitor and the
debug view at the same time, currently one turns off the other. (in the
future another window could be popped up for this ?)