InspectorGUI update + Cleanup progress
Added an OPerator Window script, that might replace the operator
monitor, it is not fully tested yet, and that is the reason that both of
them are still in. Currently only the Operator Monitor script is running
by default.
Removed several redundant scripts and elements, now the scene only
contains 3 GameObject from which only 2 is essential : MainCamera and
the PupilTracker