Recorder fix + other fixes
Finally the recorder does work with VR.
On recording a dummy camera is created under the active main camera,
that does not render to the main display nor to VR, but does see the
same thing as the main VR camera. After recording this camera is
destroyed.
The recording session is partially done on a secondary thread. The main
thread queues up the byte arrays while the secondary thread constant
loops through the queue feeding it to the FFMPG pipe.
- Some minor changes in the GUI.
The Service port is available now on Local too, because it is sometimes
not the default value.
- Now the Failed to connect error message is an actual Debug.LogWarning
not a print with the message : Failed to connect with Server. Trying
again in 5 second. So finally the retrying of connection is
implelmented.
- Automatic connecting to Service ON/OFF